Even more splitting of modules
